Initial contribution of training and info material
1. Git/Gerrit workshop
Slides and exercises for a Git/Gerrit workshop where participants learn
how to work with EGit and Gerrit.

2. Gerrit and Code Review Best Practices

3. Gerrit and Eclipse
Presentation about the integration of Gerrit in Eclipse with Mylyn and
Egit.

4. Gerrit new features
Presentations about new features in Gerrit

The slides and exercises are based on asciidoc. Slidy is used as
backend for the presentations. Graphviz is used to generate commit
graphs and other graphics.

The build result for each presentation is a self-contained html file.
Exercises can be generated as html or pdf.

All material was used in SAP, references to SAP have been removed from
the slides and have been blackened in the screenshots. The screenshots
will be renewed in another change.

Change-Id: Ia6d6724d9dee7a209e1616137d38b30dae8bb8c3